refactor: break down task executor.rs into focused modules
- Split 1501-line executor.rs into organized module hierarchy
- Largest file now 242 lines (down from 1501)
- Group by function: dependency/, execution/, runner/
- Use directory hierarchies, no underscores
- Maintain all functionality and test coverage
